handle wallet connection
