+++ catalog/src/main/java/com/getfront/catalog/ui/WebViewActivity.kt
@@ -6,7 +6,6 @@ import android.graphics.Color
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.getfront.catalog.databinding.WebviewActivityBinding
-import com.getfront.catalog.ui.web.FrontWebViewClient
 import com.getfront.catalog.utils.ifNotBlank
 import com.getfront.catalog.utils.intent
 import com.getfront.catalog.utils.viewBinding
@@ -34,7 +33,6 @@ internal class WebViewActivity : AppCompatActivity() {
         binding.webView.apply {
             settings.javaScriptEnabled = true
             settings.domStorageEnabled = true
-            webViewClient = FrontWebViewClient()
             setBackgroundColor(Color.TRANSPARENT)
             targetUrl.ifNotBlank { loadUrl(it) }
         }
