--- catalog/src/main/java/com/getfront/catalog/ui/web/FrontWebViewClient.kt
@@ -1,16 +1,0 @@
-package com.getfront.catalog.ui.web
-
-import android.webkit.WebResourceRequest
-import android.webkit.WebView
-import android.webkit.WebViewClient
-
-internal open class FrontWebViewClient : WebViewClient() {
-
-    override fun shouldOverrideUrlLoading(
-        view: WebView?,
-        request: WebResourceRequest?
-    ): Boolean {
-        // prevent some brokers from opening a browser app
-        return false
-    }
-}
